refactored internal file structure for ui
