--- internal/ui/cmds.go
@@ -1,47 +1,0 @@
-package ui
-
-import (
-	"net/http"
-	"time"
-
-	tea "github.com/charmbracelet/bubbletea"
-)
-
-const url = "https://charm.sh/"
-
-type statusMsg int
-
-type errMsg struct{ err error }
-
-func (e errMsg) Error() string { return e.err.Error() }
-
-func checkServer() tea.Msg {
-	c := &http.Client{Timeout: 10 * time.Second}
-	res, err := c.Get(url)
-
-	if err != nil {
-		return errMsg{err}
-	}
-	return statusMsg(res.StatusCode)
-}
-
-func cmdWithArg(url string) tea.Cmd {
-	return func() tea.Msg {
-		c := &http.Client{Timeout: 10 * time.Second}
-		res, err := c.Get(url)
-
-		if err != nil {
-			return errMsg{err}
-		}
-		return statusMsg(res.StatusCode)
-	}
-}
-
-func viewTracks() tea.Msg {
-
-	return nil
-}
-
-func resetState() tea.Msg {
-	return nil
-}
