+++ internal/ui/messages/messages.go
@@ -18,3 +18,7 @@ type PrevMsg bool
 
 type UserMsg *spotify.PrivateUser
 type PlaylistMsg *spotify.SimplePlaylistPage
+
+type errMsg struct{ err error }
+
+func (e errMsg) Error() string { return e.err.Error() }
