+++ internal/ui/styles/styles.go
@@ -1,35 +1,35 @@
-package ui
+package styles
 
 import "github.com/charmbracelet/lipgloss"
 
 var (
-	appStyle = lipgloss.NewStyle().Padding(1, 2)
+	AppStyle = lipgloss.NewStyle().Padding(1, 2)
 
-	menuStyle = lipgloss.NewStyle().
+	MenuStyle = lipgloss.NewStyle().
 			PaddingLeft(1).
 			PaddingRight(1).
 			BorderStyle(lipgloss.NormalBorder())
 
-	displayStyle = lipgloss.NewStyle().
-			PaddingLeft(1).
-			PaddingRight(1).
-			Background(lipgloss.Color("FFF")).
-			BorderStyle(lipgloss.NormalBorder())
-
-	viewStyle = lipgloss.NewStyle().
+	DisplayStyle = lipgloss.NewStyle().
 			PaddingLeft(1).
 			PaddingRight(1).
 			Background(lipgloss.Color("FFF")).
 			BorderStyle(lipgloss.NormalBorder())
 
-	titleStyle = lipgloss.NewStyle().
+	ViewStyle = lipgloss.NewStyle().
+			PaddingLeft(1).
+			PaddingRight(1).
+			Background(lipgloss.Color("FFF")).
+			BorderStyle(lipgloss.NormalBorder())
+
+	TitleStyle = lipgloss.NewStyle().
 			Foreground(lipgloss.Color("#FFFDF5")).
 			Background(lipgloss.Color("#25A065")).
 			Padding(0, 1)
 
-	statusMessageStyle = lipgloss.NewStyle().
+	StatusMessageStyle = lipgloss.NewStyle().
 				Foreground(lipgloss.AdaptiveColor{Light: "#04B575", Dark: "#04B575"}).
 				Render
 
-	docStyle = lipgloss.NewStyle().Margin(1, 2).Border(lipgloss.RoundedBorder())
+	DocStyle = lipgloss.NewStyle().Margin(1, 2).Border(lipgloss.RoundedBorder())
 )
